hope this works

+++ main.py
@@ -1,14 +1,20 @@
 import logging
+import os
+from dotenv import load_dotenv
 
 from aiogram.utils import executor
 
-from bot_logic.reminder_bot import dp
+from bot_logic.reminder_bot import dp, reminder_bot
+from bot_logic.utils import resume_scheduled_jobs
 from db.connection_pool import get_connection
 import db.database as database
 
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
+
+load_dotenv()
+debug = os.environ['DEBUG']
 
 
 def main():
@@ -17,13 +23,15 @@ def main():
         with get_connection() as connection:
             logger.info("No errors while connecting to the db!")
             database.create_tables(connection)
+            resume_scheduled_jobs(connection, reminder_bot, logger)
         logger.info("Starting Vladking bot...")
         executor.start_polling(dp, skip_updates=True)
     finally:
         # debug mode
-        with get_connection() as connection:
-            logger.info("Deleting tables!")
-            database.delete_tables(connection)
+        if debug == 'True':
+            with get_connection() as connection:
+                logger.info("Deleting tables!")
+                database.delete_tables(connection)
         logger.info("Stopping Vladkins bot...")
